refactoring: create seperate route for githubWebhook
